Создал новый компонент places-list.
